requirements.txt
python-dotenv==0.21.1

telebot~=0.0.5
requests~=2.28.2